lint & rename changes
