+++ deployment/deployment.json
@@ -7746,7 +7746,7 @@
           }
         },
         "files": {
-          "template": "polygon-root.template.yaml"
+          "template": "polygon.mainnet.template.yaml"
         },
         "options": {
           "prepare:yaml": true,
@@ -7768,7 +7768,7 @@
           }
         },
         "files": {
-          "template": "polygon-child.template.yaml"
+          "template": "polygon.polygon.template.yaml"
         },
         "options": {
           "prepare:yaml": true,
