--- subgraphs/polygon-bridge/.gitignore
@@ -1,3 +1,0 @@
-.DS_Store
-configure.ts
-docs/
