+++ subgraphs/polygon-bridge/src/common/constants.ts
@@ -12,4 +12,4 @@ export const crossPoolTokens = new TypedMap<
   TypedMap<BigInt, Address>
 >();
 
-export const posDecoderAddress = "0x13E301F8d9563e3D8d48F1d21aE8110B22558cd5";
+export const posDecoderAddress = "0x13e301f8d9563e3d8d48f1d21ae8110b22558cd5";
